Use containerd's apparmor package to detect if apparmor can be used
The runc/libcontainer apparmor package on master no longer checks if apparmor_parser
is enabled, or if we are running docker-in-docker.

While those checks are not relevant to runc (as it doesn't load the profile), these
checks _are_ relevant to us (and containerd). So switching to use the containerd
apparmor package, which does include the needed checks.

(Cherry-picked from 2834f84, to avoid problems with app armor in some
environments -- including balenaOS' CI.)

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
Signed-off-by: Alex Gonzalez <alexg@balena.io>