Fix "slice bounds out of range" while applying deltas
The root cause of this issue was an integer overflow in 32-bit platforms
-- specifically, when explicitly converting `io.Reader` offsets
(`int64`) to slice indices (`int`, which is only 32-bit long on 32-bit
platforms).

We already had in place the checks supposed to ensure slice bounds where
always within the expected bounds. So, in a way, this commit just
re-organizes the code so that we perform safely this required `int64` to
`int` type conversion.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch