Improve error reporting for delta base image loading
This adds some context to the error message when there's some error
loading the delta base image (for example, when the delta base image is
not found locally).

The error message we had before was generic and low-level, which didn't
give any clues about what lead to the error.

Signed-off-by: Leandro Motta Barros <leandro@balena.io>
Change-type: patch